requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
playwright>=1.40
thefuzz>=0.20
python-Levenshtein>=0.25
python-dotenv>=1.0
requests>=2.31
